Add LockerTests
